remove co, deals, and move po to component folder
