+++ client/src/components/Deals/PurchaseOrder/PurchaseOrder.css
@@ -1,0 +1,19 @@
+.crypto_logo img {
+  width: 50px;
+  margin-top: 10px;
+}
+
+#payment-divider, #paypal-button {
+  text-align: center;
+  margin-bottom: 10px;
+}
+
+.payment-form {
+  height: 400px;
+}
+
+/* z-index only works for relative position */
+.paypal-button {
+  position: relative;
+  z-index: 0 !important;
+}
