--- client/src/components/Feed/Deals/CustomizeOrder/index.js
@@ -1,1 +1,0 @@
-export { default } from "./CustomizeOrder";
