--- client/src/components/Feed/Deals/DealItem/index.js
@@ -1,1 +1,0 @@
-export { default } from "./DealItem";
