--- client/src/components/Feed/Deals/PurchaseOrder/PurchaseOrder.css
@@ -1,4 +1,0 @@
-.crypto_logo img {
-  width: 50px;
-  margin-top: 10px;
-}
